+++ src/index.html
@@ -1,16 +1,39 @@
 <!DOCTYPE html>
 <html>
   <head>
-    <title>Parcel Sandbox</title>
+    <title>ToDo(Vanilla JS)</title>
     <meta charset="UTF-8" />
   </head>
 
   <body>
-    <div id="app"></div>
-
-    <p>テストです。</p>
-    <p>反映されない…</p>
-
+    <div class="input-area">
+      <input placeholder="TODOを入力" />
+      <button>追加</button>
+    </div>
+    <div class="imcomplete-area">
+        <p class="title">未完了のTODO</p>
+        <ul>
+          <div class="list-row">
+            <li>TODOです。</li>
+            <button>完了</button>
+            <button>削除</button>
+          </div>
+          <div class="list-row">
+            <li>TODOです。</li>
+            <button>完了</button>
+            <button>削除</button>
+          </div>
+        </ul>
+    </div>
+    <div class="complete-area">
+        <p class="title">完了したTODO</p>
+        <ul>
+          <div class="list-row">
+            <li>TODOでした</li>
+            <button>戻す</button>
+          </div>
+        </ul>
+    </div>
     <script src="./index.mjs" type="module"></script>
   </body>
 </html>
